backend/requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
numpy